+++ .gitignore
@@ -1,0 +1,2 @@
+# Ignore the 3D_construction_env directory
+3D_construction_env/
